Force Gradle to use right version of spring-boot-starter-data-mongodb
gorm-monodb-spring-boot has a transitive dependency on Boot 1.0’s
spring-boot-starter-datamongodb. This isn’t a problem in Maven where
dependency management upgrades the version to match the version of Boot
that’s used. Unfortunately, Gradle doesn’t have the same concept of
dependency management and, as a result, the old version of
spring-boot-starter-datamongodb is used by default. To overcome this
an explicit runtime dependency on spring-boot-starter-datamongodb is
declared. This dependency is sufficient to cause Gradle to use the
right version.